Splitting resource components for resource, conceptual models.
The pages relating to the resource and the conceptual components of the model
have been split out into seperate subdirectories. I updated the listing of
models in layouts.erb to accomodate this change, as well as changing this
file to accomodate more than just the allele model (this change is a little
ahead of time, but it feels better to do it now). Each model must be listed
explicitly by name in /data/models.yml